Merged all Updated Pull Requests
Latest "Working" Copy of the code including new GPS code (run over
UART2) and V-Tail Mixing Code

- Errors with this exist with the GPS code where sometimes the GPS
doesn't sync properly and outputs weird values. GPS data is fine until
the plane is hand launched for some reason. We have yet o be able to
replicate this on the ground. At this point data is still received,
except apparently we're at the equator. As well, I believe that the
Altimeter data is messed up once this happens, but not fully certain. A
power cycle of the Autopilot fixes this.

- The V-Tail mixin also has a trim offset that changes the V-Tail's
position when Steve switches from Autopilot to Manual. Meaning that trim
is inconsistent between modes, which could throw Steve off during a
recovery attempt

- The plane has flown with this code. No new path following code is
added, but unsure with bench testing (ground parking lot testing) if the
path following code is working correctly