Altimeter Issue
Once the plane gets to a certain height, the gps and altimeter data gets
corrupted. We've replicated this on the ground by tricking the altimeter
into flying higher (low pressure in a bag). changing the gps lattitude
and longitude data type to long double appears to fix it. Not sure if
I'm just covering up an issue and not really solving the problem
though....