Added a way to orbit the plane around a specified location
The added lines are the exact same logic as used elsewhere in the code
(copy and pasted into the if statement to keep the order of calculations
the same as before outside the if statement). The followOrbit function
is called here as it simply causes the plane to follow an orbit
independant of an entrance and exit point. To cause the plane to exit
the loop the target waypoint would have to be updated or destroyed. The
size of the orbit can be chaged using the waypoint radius value.